+++ src/aioway/chunks/chunks/exprs.py
@@ -11,13 +11,13 @@ from torch import Tensor
 
 from aioway import _typing
 from aioway._errors import GitHubTicketFiled
-from aioway._previews import AttrSet
 from aioway._tensor_exprs import (
     BatchTensorDictExpr,
     ItemTensorDictExpr,
     RenameTensorDictExpr,
     TensorDictExpr,
 )
+from aioway.tds import AttrSet
 
 from ..vectors import Vector, VectorExpr
 from .chunks import Chunk
