+++ src/aioway/chunks/vectors/exprs.py
@@ -7,10 +7,10 @@ from typing import Self
 
 from torch import Tensor
 
-from aioway._previews import Attr, AttrTerm
 from aioway._tensor_exprs import SourceTensorExpr, TensorExpr
 from aioway._tracking import logging
 from aioway._typing import AnyUFunc1, AnyUFunc2
+from aioway.tensors import Attr, AttrTerm
 
 from .vectors import Vector, VectorRhs
