+++ src/aioway/datasets/streams/sources.py
@@ -14,9 +14,9 @@ from typing import Self
 from torch.utils.data import DataLoader, Sampler
 
 from aioway import _typing
-from aioway._previews import AttrSet
 from aioway._tracking import logging
 from aioway.chunks import Chunk
+from aioway.tds import AttrSet
 
 from ..frames import Frame
 from .streams import Stream
