+++ src/aioway/tds/__init__.py
@@ -2,7 +2,4 @@
 
 from . import _validation
 from .attrs import *
-from .devices import *
-from .dtypes import *
-from .sets import *
-from .shapes import *
+from .fn import *
