+++ src/aioway/tensors/attrs/README.md
