+++ src/aioway/tensors/attrs/__init__.py
@@ -1,4 +1,6 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 from .attrs import *
-from .fn import *
+from .devices import *
+from .dtypes import *
+from .shapes import *
