+++ src/aioway/tensors/attrs/attrs.py
