+++ src/aioway/tensors/attrs/devices.py
