+++ src/aioway/tensors/attrs/shapes.py
