+++ tests/attrs/test_sets.py
@@ -6,9 +6,9 @@ import torch
 from pytest import FixtureRequest
 from tensordict import TensorDict
 
-from aioway import _previews
-from aioway._previews import Attr, AttrSet, _validation
 from aioway.chunks import Chunk
+from aioway.tds import AttrSet, _validation
+from aioway.tensors import Attr
 
 
 @pytest.fixture
@@ -76,7 +76,7 @@ def test_validation_ok(schema: AttrSet, valid_data: TensorDict) -> None:
 
 def test_construction_of_attrset(valid_data: TensorDict):
     parsed = AttrSet.from_tensordict(valid_data)
-    assert parsed == _previews.attr_set(
+    assert parsed == AttrSet.parse(
         {
             "a": Attr.parse(device="cpu", shape=[11, 2, 3], dtype="int32"),
             "b": Attr.parse(device="cpu", shape=[11, 6], dtype="float32"),
