+++ tests/datasets/streams/test_maps.py
@@ -7,7 +7,6 @@ from collections.abc import Callable
 import pytest
 from pytest import FixtureRequest
 
-from aioway._previews import AttrSet
 from aioway.chunks import Chunk
 from aioway.datasets import (
     ApplyStream,
@@ -19,6 +18,7 @@ from aioway.datasets import (
     Stream,
     StreamState,
 )
+from aioway.tds import AttrSet
 
 
 @dcls.dataclass
